Title: 낮은 지점들, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240116/낮은 지점들/lowest-points.cpp
@@ -3,11 +3,11 @@
 using namespace std;
 
 int main() {
-    unordered_map<long, long> m;
+    unordered_map<long long, long long> m;
     int n;
     cin >> n;
 
-    long key, value;
+    long long key, value;
     for (int i = 0; i < 5; ++i)
     {
         cin >> key >> value;
